sparse_mor/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
pytest>=7.3.1
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.2.0  # for evaluation metrics
tensorboard>=2.13.0  # for training visualization
typing-extensions>=4.5.0
accelerate>=0.20.0  # for optimized transformer operations
plotly>=5.13.0      # for interactive visualizations
kaleido>=0.2.1      # for static image export
pandas>=1.5.0       # for data manipulation